+++ internal/dnsrecordresourcevalidator/hostname_test.go
@@ -44,6 +44,41 @@ func TestHostname(t *testing.T) {
 				"value": tftypes.NewValue(tftypes.String, "some-value."),
 			},
 		},
+		{
+			ExpectedError: false,
+			PlanValues: map[string]tftypes.Value{
+				"type":  tftypes.NewValue(tftypes.String, "MX"),
+				"value": tftypes.NewValue(tftypes.String, "mail.example.com"),
+			},
+		},
+		{
+			ExpectedError: true,
+			PlanValues: map[string]tftypes.Value{
+				"type":  tftypes.NewValue(tftypes.String, "MX"),
+				"value": tftypes.NewValue(tftypes.String, "mail.example.com."),
+			},
+		},
+		{
+			ExpectedError: false,
+			PlanValues: map[string]tftypes.Value{
+				"type":  tftypes.NewValue(tftypes.String, "MX"),
+				"value": tftypes.NewValue(tftypes.String, "."),
+			},
+		},
+		{
+			ExpectedError: false,
+			PlanValues: map[string]tftypes.Value{
+				"type":  tftypes.NewValue(tftypes.String, "TXT"),
+				"value": tftypes.NewValue(tftypes.String, "."),
+			},
+		},
+		{
+			ExpectedError: true,
+			PlanValues: map[string]tftypes.Value{
+				"type":  tftypes.NewValue(tftypes.String, "CNAME"),
+				"value": tftypes.NewValue(tftypes.String, "."),
+			},
+		},
 	}
 
 	configSchema := schema.Schema{